Create Impressum page
The AI will create a new "Impressum" page based on the provided contact information and standard German legal requirements. It will also update the footer component to reflect the new "Impressum" page and remove any redundant contact information from it. The AI will extract the following information for the Impressum page:

**Contact Information:**
*   Magnus Zastrow
    *   Ansprechpartner für Anmeldung, Sponsoring & allgemeine Fragen
    *   Tel.: +49 176 363 878 11
    *   E-Mail: spendenlauf.lueneburg@gmail.com
*   Damien Massenbach
    *   Ansprechpartner Logistik & Ablauf
    *   Tel.: +49 176 442 272 23
*   Benedikt Löbbecke
    *   Ansprechpartner Social- & Print-Media
    *   Tel.: +49 151 419 986 21
*   Anschrift:
    *   Lt Zastrow
    *   5./ AufklLehrBtl 3 Lüneburg
    *   Bleckeder Landstraße 59
    *   21337 Lüneburg

The AI will also ensure that the footer is updated to include a link to the new "Impressum" page.